+++ ChangeLog.md
@@ -5,6 +5,7 @@ IMPROVEMENTS
 
 BUG FIXES
 - fix [390](https://github.com/AlexAkulov/clickhouse-backup/issues/390), respect skip_tables patterns during restore and skip all INFORMATION_SCHEMA related tables even skip_tables don't contain INFORMATION_SCHEMA pattern
+- fix [388](https://github.com/AlexAkulov/clickhouse-backup/issues/388), improve restore ON CLUSTER for VIEW with TO clause
 
 # v1.3.0